requirements.txt
pillow>=10.0.0
transformers>=4.44.0
diffusers>=0.30.0
safetensors>=0.4.2
accelerate>=0.33.0
huggingface_hub>=0.24.0
tk